+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## unreleased
+
+* Update Twilio Conversations Android SDK to [6.0.3](https://www.twilio.com/docs/conversations/android/changelog#conversations-603-october-26-2023)
+* Update Twilio Conversations iOS SDK to [4.0.2](https://www.twilio.com/docs/conversations/ios/changelog#conversations-402-august-3-2023)
+
 ## 1.1.6+8  
 * Fixed issue [#21](https://github.com/Diversido/flutter_twilio_conversations/issues/21)